[WIP] Moved business logic of looking up creds from IAM Role in separate method
Also added a new method `getKeysFromIamTaskRole(env)` that is responsible for checking if it can retrieve the keys in case neither an access key nor an IAM role have been defined in the config. To do this, it looks up the special environment variable provided by the ECS agent, when running in an ECS node.