Removed checks for what AWS config is valid
The logic of what configuration is valid (access + secret key vs. iam role vs. neither, which means IAM task role, if env var is set) is all implemented in the `DescribeInstances` class, so if we check these here too, it will mean duplication of the validation logic. Leave it up to `DescribeInstances` and its tests to verify that.