Added test that checks the IAM role configuration case
Check that when iam role has been set and no access key has been set, we can correctly retrieve and parse the credentials.
In order to keep this as a Unit Test, we needed to mock out the HTTP calls, as well as their responses, so this test only covers the business logic on the library side. The UT does not cover testing that the EC2 instance actually does have the IAM role (nor should it, imho).