Added test that checks the IAM Task role
Check that when NO IAM role or Access Key have been set in the config, we can correctly retrieve and parse the credentials IF the appropriate env var has been set by the ECS agent.
In order to keep this as a Unit Test, we needed to mock out the System.getenv calls, which we did through the usage of the Environment wrapper class.